+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/application/PersonRestController.java
@@ -3,16 +3,13 @@ package com.nimio2024.nimio2024_tfg_polmorenocastelar.application;
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.application.dto.PersonDTO;
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.controllers.PersonController;
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.domain.Person;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping("/persons")
 public class PersonRestController {
 
-    PersonController    personController;
+    PersonController  personController;
 
     public PersonRestController(PersonController personController) {
         this.personController = personController;
@@ -21,6 +18,11 @@ public class PersonRestController {
     @PostMapping("")
     public Person createPerson(@RequestBody PersonDTO personDTO) {
         return personController.createPerson(personDTO);
+    }
+
+    @DeleteMapping("/{personId}")
+    public void deletePerson(@PathVariable Long personId) {
+        personController.deletePerson(personId);
     }
 
 }