+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/controllers/PersonController.java
@@ -19,4 +19,8 @@ public class PersonController {
         Person person = new Person(personDTO);
         return personService.savePerson(person);
     }
+
+    public void deletePerson(Long personId) {
+        personService.deletePerson(personId);
+    }
 }