+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/domain/Collector.java
@@ -76,5 +76,4 @@ public class Collector {
         this.person = person;
     }
 
-
 }