+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/domain/Person.java
@@ -75,4 +75,7 @@ public class Person {
     public void setCollectorList(List<Collector> collectorList) {
         this.collectorList = collectorList;
     }
+
+
+
 }