+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/domain/Student.java
@@ -93,4 +93,13 @@ public class Student {
 
         this.collectorList.add(collector);
     }
+
+    public void deleteAuth(Person person) {
+        for (Collector collector : collectorList) {
+            if (collector.getPerson().equals(person)) {
+                collectorList.remove(collector);
+                break;
+            }
+        }
+    }
 }