+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/persistence/CollectorRepository.java
@@ -1,9 +1,15 @@
 package com.nimio2024.nimio2024_tfg_polmorenocastelar.persistence;
 
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.domain.Collector;
+import com.nimio2024.nimio2024_tfg_polmorenocastelar.domain.Person;
+import com.nimio2024.nimio2024_tfg_polmorenocastelar.domain.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
 public interface CollectorRepository extends JpaRepository<Collector, Long> {
+
+    @Query("SELECT c FROM Collector c WHERE c.student = ?1 AND c.person = ?2")
+    Collector findCollectorByStudentAndPerson(Student student, Person person);
 }