+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/service/CollectorService.java
@@ -1,6 +1,8 @@
 package com.nimio2024.nimio2024_tfg_polmorenocastelar.service;
 
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.domain.Collector;
+import com.nimio2024.nimio2024_tfg_polmorenocastelar.domain.Person;
+import com.nimio2024.nimio2024_tfg_polmorenocastelar.domain.Student;
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.persistence.CollectorRepository;
 import org.springframework.stereotype.Service;
 
@@ -14,5 +16,9 @@ public class CollectorService {
 
     public void saveCollector(Collector collector) {
         collectorRepository.save(collector);
+    }
+
+    public void deleteCollectorByStudentAndPerson(Student student, Person person) {
+        collectorRepository.delete(collectorRepository.findCollectorByStudentAndPerson(student, person));
     }
 }