+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/service/PersonService.java
@@ -19,4 +19,8 @@ public class PersonService {
     public Person getPersonById(Long personId) {
         return personRepository.findById(personId).orElse(null);
     }
+
+    public void deletePerson(Long personId) {
+        personRepository.deleteById(personId);
+    }
 }