+++ NIMIO2024_TFG_PolMorenoCastelar/changelog.md
@@ -82,6 +82,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - [NPP-44] 》ST38F - Create a student associated with a course
 - [NPP-55] 》ST46F - Create a student without associated course
 - [NPP-45] 》ST39F - Update student by ID
+- [NPP-47] 》ST41F - Delete Auth without deleting the person
 - [NPP-48] 》ST42F - Get students by courseId
 - [NPP-49] 》ST43F - Get student by ID
 - [NPP-54] 》ST48F - Delete a student by ID